Change place of two recycler view. Bottom - event , left - weather at week.
