+++ app/src/main/java/com/example/mac/suchik/UI/MainWindowFragment.java
@@ -192,8 +192,8 @@ public class MainWindowFragment extends Fragment implements Callbacks, AdapterVi
         imagePressure = view.findViewById(R.id.image_pressure);
         rv_event = view.findViewById(R.id.rv_event);
         rv_clothes.setLayoutManager(new LinearLayoutManager(getContext(), LinearLayoutManager.VERTICAL, false));
-        rv_event.setLayoutManager(new LinearLayoutManager(getContext() , LinearLayoutManager.VERTICAL , false));
-        rv.setLayoutManager(new LinearLayoutManager(getContext(), LinearLayoutManager.HORIZONTAL, false));
+        rv_event.setLayoutManager(new LinearLayoutManager(getContext() , LinearLayoutManager.HORIZONTAL , false));
+        rv.setLayoutManager(new LinearLayoutManager(getContext(), LinearLayoutManager.VERTICAL, false));
         List<String> data = new ArrayList<>();
 
         DateFormat dateFormat = new SimpleDateFormat("dd.MM.yyyy", Locale.getDefault());
