+++ app/src/main/res/layout/for_list_weather.xml
@@ -3,20 +3,8 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/for_list_weather"
-    android:layout_width="100dp"
-    android:layout_height="150dp">
-
-    <TextView
-        android:id="@+id/for_main"
-        android:layout_width="80dp"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_marginTop="3dp"
-        android:layout_marginEnd="10dp"
-        android:gravity="center"
-        android:text="Четверг"
-        android:textAlignment="center"
-        android:textSize="20dp" />
+    android:layout_width="match_parent"
+    android:layout_height="50dp">
 
     <!--<View-->
     <!--android:id="@+id/line"-->
@@ -25,26 +13,33 @@
     <!--android:background="#000"-->
     <!--android:layout_marginTop="3dp"-->
     <!--android:layout_below="@+id/part_of_sun"/>-->
+    <ImageView
+        android:id="@+id/part_of_sun"
+        android:layout_width="30dp"
+        android:layout_height="30dp"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentBottom="true"
+
+        android:layout_marginStart="0dp"
+        android:layout_marginBottom="14dp"
+        android:layerType="software"
+        android:src="@drawable/sunny" />
+
     <TextView
         android:id="@+id/temp_avg"
         android:layout_width="80dp"
         android:layout_height="40dp"
-        android:layout_below="@+id/part_of_sun"
-        android:layout_alignParentEnd="true"
-        android:layout_marginEnd="10dp"
-        android:layout_marginBottom="3dp"
+        android:layout_toRightOf="@+id/part_of_sun"
         android:gravity="center"
         android:textAlignment="center"
         android:textSize="20sp" />
 
-    <ImageView
-        android:id="@+id/part_of_sun"
+    <TextView
+        android:id="@+id/for_main"
         android:layout_width="80dp"
-        android:layout_height="80dp"
-        android:layout_below="@+id/for_main"
-        android:layout_alignParentEnd="true"
-        android:layout_marginEnd="10dp"
-        android:layerType="software"
-        android:src="@drawable/sunny" />
+        android:layout_height="wrap_content"
+        android:layout_toRightOf="@id/temp_avg"
+        android:gravity="center"
+        android:text="Четверг" />
 
 </RelativeLayout>
